Feat/7 refresh token

See merge request s11-bigdata-recom-sub1/S11P21B205!16

+++ Plog/src/main/java/com/plog/backend/domain/member/repository/MemberImageRepository.java
@@ -4,7 +4,7 @@ import com.plog.backend.domain.member.entity.Member;
 import com.plog.backend.domain.member.entity.MemberImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberImageRepository extends JpaRepository<MemberImage, Integer> {
+public interface MemberImageRepository extends JpaRepository<MemberImage, Long> {
 
     MemberImage findByMember(Member member);
 }
